refactor: extract createToken logic from issueHandler
