+++ src/functions/issueHandler.ts
@@ -6,10 +6,9 @@ import {
 import { logger } from "../logging/logger";
 import { LogMessage } from "../logging/LogMessage";
 import { randomUUID } from "crypto";
-import { sign } from "../common/aws/kms";
 import { putObject } from "../common/aws/s3";
-import format from "ecdsa-sig-formatter";
 import { getConfig } from "../config/getConfig";
+import { createToken, StatusList } from "../../test/common/token/createToken";
 
 const REQUIRED_ENV_VARS = [
   "SIGNING_KEY_ID",
@@ -21,14 +20,6 @@ interface Configuration {
   index: number;
   statusList: StatusList;
 }
-
-interface StatusList {
-  bits: number;
-  lst: string;
-}
-
-const TTL = 2592000;
-const ALGORITHM = "ES256";
 
 export async function handler(
   _event: APIGatewayProxyEvent,
@@ -73,48 +64,5 @@ function getRandomConfig(): Configuration {
     },
   ];
 
-  // NOSONAR: Using Math.random() is safe here as security-critical randomness is not required
-  return configurations[Math.floor(Math.random() * configurations.length)];
-}
-
-export async function createToken(
-  statusList: StatusList,
-  uri: string,
-  keyId: string,
-) {
-  const header = buildHeader(keyId);
-  const payload = buildPayload(statusList, uri);
-
-  const encodedHeader = base64Encoder(JSON.stringify(header));
-  const encodedPayload = base64Encoder(JSON.stringify(payload));
-
-  const message = `${encodedHeader}.${encodedPayload}`;
-
-  const encodedSignature = base64Encoder(await sign(message, keyId));
-  const signature = format.derToJose(encodedSignature, ALGORITHM);
-
-  return `${message}.${signature}`;
-}
-
-function buildHeader(keyId: string) {
-  return {
-    alg: ALGORITHM,
-    kid: keyId,
-    typ: "statuslist+jwt",
-  };
-}
-
-function buildPayload(statusList: StatusList, uri: string) {
-  const timestamp = Math.floor(Date.now() / 1000);
-  return {
-    iat: timestamp,
-    exp: timestamp + TTL,
-    status_list: statusList,
-    sub: uri,
-    ttl: TTL,
-  };
-}
-
-function base64Encoder(data: string | Uint8Array<ArrayBufferLike>) {
-  return Buffer.from(data).toString("base64url");
+  return configurations[Math.floor(Math.random() * configurations.length)]; // NOSONAR: Using Math.random() is safe here as security-critical randomness is not required
 }
